#student: Dodata lista predmeta koje slusaju studenti

+++ src/gui/model/ListaStudenata.java
@@ -83,7 +83,7 @@ public class ListaStudenata implements Serializable{
 	}
 	
 	public void dodajStudenta(String ime, String prezime, String datumRodjenja, String adresa, String telefon, String eMail,
-			String brIndeksa, String datumUpisa, int godStudija, NacinFinansiranja finansiranje, double prosecnaOcena)
+			String brIndeksa, String datumUpisa, short godStudija, NacinFinansiranja finansiranje, double prosecnaOcena)
 	{
 		this.studenti.add(new Student(ime, prezime, datumRodjenja, adresa, telefon, eMail, brIndeksa, datumUpisa, godStudija, finansiranje, prosecnaOcena));
 	}
@@ -101,7 +101,7 @@ public class ListaStudenata implements Serializable{
 	}
 	
 	public void izmeniStudenta(String brI, String ime, String prezime, String datumRodjenja, String adresa, String telefon, String eMail,
-			String datumUpisa, int godStudija, NacinFinansiranja finansiranje, double prosecnaOcena)
+			String datumUpisa, short godStudija, NacinFinansiranja finansiranje, double prosecnaOcena)
 	{
 		for(Student s : studenti)
 		{
